Move subclasses-related stuff to new subclasses.{h, cc}
